refactor: simplify E2E testing setup
- Move kitchen-sink template to examples/e2e-studio with pnpm link overrides
- Move playwright tests to packages/playground/e2e/
- Support both port 5173 (vite dev) and 4111 (mastra dev)
- Remove Verdaccio local registry setup
- Remove changeset snapshot versioning infrastructure
