--- e2e-tests/_local-registry-setup/index.js
@@ -1,3 +1,0 @@
-export * from './registry.js';
-export * from './publish.js';
-export * from './prepare.js';
