--- e2e-tests/_local-registry-setup/package.json
@@ -1,8 +1,0 @@
-{
-  "name": "local-registry-setup",
-  "version": "0.0.1",
-  "private": true,
-  "main": "index.js",
-  "type": "module",
-  "dependencies": {}
-}
