--- e2e-tests/_local-registry-setup/verdaccio.yaml
@@ -1,47 +1,0 @@
-storage: ./storage
-
-# Increase max body size to handle large packages
-max_body_size: 100mb
-
-uplinks:
-  npmjs:
-    url: https://registry.npmjs.org/
-
-packages:
-  # Block npm fallback for @mastra/* packages - they MUST be published locally
-  # This ensures e2e tests fail fast if a package is missing from the publish list
-  '@mastra/*':
-    access: $all
-    publish: $all
-    unpublish: $all
-    # No proxy - will 404 if not published locally
-
-  # Same for the unscoped 'mastra' package
-  'mastra':
-    access: $all
-    publish: $all
-    unpublish: $all
-    # No proxy - will 404 if not published locally
-
-  # All other packages can fall back to npm
-  '**':
-    access: $all
-    publish: $all
-    unpublish: $all
-    proxy: npmjs
-
-# Security settings to allow everything
-security:
-  api:
-    legacy: true
-
-# Disable audit
-middlewares:
-  audit:
-    enabled: false
-
-web:
-  enable: true
-
-log:
-  - { type: stdout, format: pretty, level: error }
