--- e2e-tests/kitchen-sink/.gitignore
@@ -1,7 +1,0 @@
-
-# Playwright
-node_modules/
-/test-results/
-/playwright-report/
-/blob-report/
-/playwright/.cache/
