--- e2e-tests/kitchen-sink/package.json
@@ -1,39 +1,0 @@
-{
-  "name": "e2e-kitchen-sink",
-  "version": "0.0.0",
-  "private": true,
-  "type": "module",
-  "description": "E2E tests for @mastra/create-mastra",
-  "scripts": {
-    "test": "vitest run",
-    "test:e2e": "playwright test",
-    "test:e2e:ui": "playwright test --ui",
-    "test:e2e:setup": "playwright install"
-  },
-  "devDependencies": {
-    "@playwright/test": "^1.56.0",
-    "@types/node": "22.13.17",
-    "@types/react": "^19.2.2",
-    "@types/react-dom": "^19.2.1",
-    "execa": "^9.6.0",
-    "get-port": "^7.1.0",
-    "globby": "^14.1.0",
-    "playwright": "^1.56.0",
-    "tsx": "^4.20.3",
-    "verdaccio": "^6.2.0",
-    "verdaccio-auth-memory": "^10.3.1",
-    "vitest": "^4.0.8"
-  },
-  "dependencies": {
-    "nanoid": "^5.1.6",
-    "npm-cli-login": "^1.0.0",
-    "playwright": "^1.56.0",
-    "react": "^19.2.0",
-    "react-dom": "^19.2.0",
-    "tsx": "^4.20.6",
-    "verdaccio": "^6.2.0",
-    "verdaccio-auth-memory": "^10.3.1",
-    "vite": "^7.3.0",
-    "vitest": "^3.2.4"
-  }
-}
