--- e2e-tests/kitchen-sink/template/package.json
@@ -1,31 +1,0 @@
-{
-  "name": "kitchen-sink",
-  "version": "1.0.0",
-  "main": "index.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "dev": "mastra dev",
-    "build": "mastra build",
-    "start": "mastra start"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "Apache-2.0",
-  "description": "",
-  "packageManager": "pnpm@10.18.2",
-  "dependencies": {
-    "ai": "5.1.0-beta.13",
-    "mastra": "kitchen-sink-e2e-test",
-    "@ai-sdk/openai": "^2.0.44",
-    "@mastra/core": "kitchen-sink-e2e-test",
-    "@mastra/loggers": "kitchen-sink-e2e-test",
-    "@mastra/memory": "kitchen-sink-e2e-test",
-    "@mastra/libsql": "kitchen-sink-e2e-test",
-    "@mastra/mcp": "kitchen-sink-e2e-test",
-    "zod": "^3.25.76"
-  },
-  "devDependencies": {
-    "@types/node": "22.13.17",
-    "typescript": "^5.9.3"
-  }
-}
