+++ examples/e2e-studio/fixtures/text-stream.fixture.ts
