+++ examples/e2e-studio/fixtures/tool-stream.fixture.ts
