+++ examples/e2e-studio/fixtures/workflow-stream.fixture.ts
