+++ examples/e2e-studio/src/mastra/mcps/index.ts
