+++ examples/e2e-studio/src/mastra/tools/index.ts
