+++ examples/e2e-studio/tsconfig.json
