+++ packages/playground/e2e/tests/agents/$agentId/page.spec.ts-snapshots/agent-panels-overview-general-information-1.aria.yml
