+++ packages/playground/e2e/tests/agents/$agentId/page.spec.ts-snapshots/overall-layout-information-1.aria.yml
