+++ packages/playground/e2e/tests/mcps/page.spec.ts-snapshots/has-overall-information-1.aria.yml
