+++ packages/playground/e2e/tests/workflows/$workflowId/page.spec.ts-snapshots/overall-layout-information-1.aria.yml
